Fix bad close on alt song
When you picked an alternate for a given song, then closed the song
details, the list was not properly scrolled to the top of the screen
because the alt slug is not anywhere in the list. Adding the `safeSlug`,
which is the slug for the non-alternate version of the song, gives us
access to the proper `id`.